+++ guiqwt/__init__.py
@@ -23,7 +23,7 @@ External resources:
 """
 
 
-__version__ = "4.4.1"  # Update here *AND* in setup.py!
+__version__ = "4.5.0"  # Update here *AND* in setup.py!
 # (Until setup.py has been fully retrofitted, this manual sync is mandatory)