+++ guiqwt/tests/cross_section.py
@@ -13,6 +13,7 @@ import os.path as osp, numpy as np
 
 from guiqwt.plot import ImageDialog
 from guiqwt.builder import make
+from guiqwt.styles import LUTAlpha
 
 
 def create_window():
@@ -37,7 +38,7 @@ def test():
     win = create_window()
     image = make.image(filename=filename, colormap="bone")
     data2 = np.array(image.data.T[200:], copy=True)
-    image2 = make.image(data2, title="Modified", alpha_mask=True)
+    image2 = make.image(data2, title="Modified", alpha_function=LUTAlpha.LINEAR)
     plot = win.get_plot()
     plot.add_item(image)
     plot.add_item(image2, z=1)