+++ guiqwt/tools.py
@@ -1502,6 +1502,10 @@ class CommandTool(GuiTool):
         if plot is not None:
             self.activate_command(plot, checked)
 
+    def activate_command(self, plot, checked):
+        """Activate tool"""
+        pass
+
     def set_status_active_item(self, plot):
         item = plot.get_active_item()
         if item: